change meditation table schema

+++ data/import_meditation.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS `meditations` (
   `id` int(11) NOT NULL AUTO_INCREMENT,
   `time` bigint(20) NOT NULL,
-  `duration` int(11) NOT NULL,
+  `duration` bigint(20) NOT NULL,
   PRIMARY KEY (`id`)
 ) DEFAULT CHARSET=utf8 AUTO_INCREMENT=5 ;
